+++ CHANGELOG.md
@@ -8,3 +8,4 @@ All notable changes to this project will be documented in this file.
 - PHPUnit testing
 - Router facade
 - Controller invoking functionality
+- Getting router via DI container